Problema resolvido

+++ custom/php/edicao-de-dados.php
@@ -1,6 +1,6 @@
 <?php
 require_once("custom/php/common.php");
-//echo "mudou4";
+echo "mudou4";
 //ESTEBELECE LIGAÇÃO COM A BASE DE DADOS:
 $mySQL = ligacaoBD();
 
@@ -166,6 +166,9 @@ if (!mysqli_select_db($mySQL, "bitnami_wordpress")) {
                         echo "<input type='radio' id='tipo_valor' value='" . $val_tip . "' name='tipo_valor'><span class='textoLabels'>" . $val_tip . "</span><br>";
                     }
                 }
+
+                $_SESSION["value_type_subitem"] = $linhaSubitem["value_type"];
+
 
                 //ESCOLHA DO TIPO DE ITEM:
                 echo "<br><strong>Item<span class='warning'>*</span>:</strong></br>";
@@ -567,15 +570,20 @@ if (!mysqli_select_db($mySQL, "bitnami_wordpress")) {
                     //MENSAGEM DE ERRO NO CÓDIGO SQL
                     echo "<span class='warning'>Erro: " . $insertQuery . "<br>" . mysqli_error($mySQL) . "</span>";
                 } else {
-                    $deleteQuery = "DELETE FROM value WHERE value IN (SELECT value FROM subitem_allowed_value WHERE subitem_id ='" . $_SESSION["id"] . "')";
-                    if (!mysqli_query($mySQL, $deleteQuery)) {
-                        //MENSAGEM DE ERRO NO CÓDIGO SQL
-                        echo "<span class='warning'>Erro: " . $deleteQuery . "<br>" . mysqli_error($mySQL) . "</span>";
-                    }
-                    $deleteQuery = "DELETE FROM subitem_allowed_value WHERE subitem_id ='" . $_SESSION["id"] . "'";
-                    if (!mysqli_query($mySQL, $deleteQuery)) {
-                        //MENSAGEM DE ERRO NO CÓDIGO SQL
-                        echo "<span class='warning'>Erro: " . $deleteQuery . "<br>" . mysqli_error($mySQL) . "</span>";
+                    //SE O VALUE_TYPE ERA "ENUM" E FOI ALTERADO:
+                    if($_SESSION["value_type_subitem"] == "enum" && $tipo_valor != "enum") {
+                        //QUERY PARA APAGAR TODOS OS TUPLOS DA TABELA value DESSE SUBITEM:
+                        $deleteQuery = "DELETE FROM value WHERE value IN (SELECT value FROM subitem_allowed_value WHERE subitem_id ='" . $_SESSION["id"] . "')";
+                        if (!mysqli_query($mySQL, $deleteQuery)) {
+                            //MENSAGEM DE ERRO NO CÓDIGO SQL
+                            echo "<span class='warning'>Erro: " . $deleteQuery . "<br>" . mysqli_error($mySQL) . "</span>";
+                        }
+                        //QUERY PARA APAGAR TODOS OS TUPLOS DA TABELA subitem_allowed_value DESSE SUBITEM:
+                        $deleteQuery = "DELETE FROM subitem_allowed_value WHERE subitem_id ='" . $_SESSION["id"] . "'";
+                        if (!mysqli_query($mySQL, $deleteQuery)) {
+                            //MENSAGEM DE ERRO NO CÓDIGO SQL
+                            echo "<span class='warning'>Erro: " . $deleteQuery . "<br>" . mysqli_error($mySQL) . "</span>";
+                        }
                     }
                     //O UTILIZADOR É INFORMADO SOBRE O SUCESSO DA EDIÇÃO:
                     echo "<span class='information'>Alterou os dados do subitem com sucesso.<br>Clique em <strong>Continuar</strong> para avançar.<br></span>";
